% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ChromoCorrect-package.R
\docType{package}
\name{ChromoCorrect-package}
\alias{ChromoCorrect}
\alias{ChromoCorrect-package}
\title{ChromoCorrect: Detection and Correction of Chromosomal Location Bias in Read Counts}
\description{
A package and R Shiny app that detects and corrects chromosomal location bias in read count data.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/gerisullivan/ChromoCorrect}
  \item Report bugs at \url{https://github.com/gerisullivan/ChromoCorrect/issues}
}

}
\author{
\strong{Maintainer}: Geraldine Sullivan \email{geraldine.sullivan@hdr.mq.edu.au} (\href{https://orcid.org/0000000152538660}{ORCID})

}
\keyword{internal}
