% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detect_bias.R
\name{detect_bias}
\alias{detect_bias}
\title{detect_bias}
\usage{
detect_bias(path = "/logfcs", locusInfo = TRUE, savePlot = TRUE)
}
\arguments{
\item{path}{= "/logFCs"}

\item{locusInfo}{= TRUE}

\item{savePlot}{= TRUE}
}
\description{
Generate a png of locus by fold change scatterplot for each output file in a directory to determine whether any are affected by chromosomal location bias.
}
