% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/launch_app.R
\name{launch_app}
\alias{launch_app}
\title{Launch Shiny App}
\usage{
launch_app(name = "app.R")
}
\arguments{
\item{name}{The name of the app to run}

\item{...}{arguments to pass to shiny::runApp}
}
\description{
Launch Shiny App
}
