% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalise_bias.R
\name{normalise_bias}
\alias{normalise_bias}
\title{normalise_bias}
\usage{
normalise_bias(
  x,
  windowSize = "auto",
  minrc = 10,
  control = "",
  writePlots = TRUE,
  locusInfo = TRUE,
  path = getwd()
)
}
\arguments{
\item{x}{= readcounts}

\item{windowSize}{= "auto"}

\item{minrc}{= 10}

\item{control}{= ""}

\item{writePlots}{= TRUE}

\item{locusInfo}{= TRUE}

\item{path}{= getwd()}
}
\description{
normalise_bias
}
