% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/structure_rc.R
\name{structure_rc}
\alias{structure_rc}
\title{structure_rc}
\usage{
structure_rc(
  csvpath = "./readcounts",
  getLocusInfo = TRUE,
  suffix = ".tradis_gene_insert_sites.csv"
)
}
\arguments{
\item{csvpath}{"./readcounts"}

\item{getLocusInfo}{= TRUE}

\item{suffix}{= ".tradis_gene_insert_sites.csv"}
}
\description{
structure_rc
}
