% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detect_bias.R
\name{detect_bias}
\alias{detect_bias}
\title{detect_bias()}
\usage{
detect_bias(path = "/logfcs", locusInfo = TRUE, savePlot = TRUE)
}
\arguments{
\item{path}{= "/logfcs"}

\item{locusInfo}{= TRUE}

\item{savePlot}{= TRUE}
}
\description{
detect_bias()
}
